implementado  crud Clientes

+++ reservas-hotel/src/main/java/com/reservashotel/model/repository/ClienteRespository.java
@@ -4,6 +4,6 @@ package com.reservashotel.model.repository;
 import com.reservashotel.model.entities.ClienteEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteRespository  extends JpaRepository<ClienteEntity, Integer>{
+public interface ClienteRespository  extends JpaRepository<ClienteEntity, Long>{
 
 }
