+++ reservas-hotel/src/main/java/com/reservashotel/model/repository/HotelRespository.java
@@ -4,5 +4,5 @@ package com.reservashotel.model.repository;
 import com.reservashotel.model.entities.HotelEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HotelRespository extends JpaRepository<HotelEntity, Integer> {
+public interface HotelRespository extends JpaRepository<HotelEntity, Long> {
 }
